==> 3_dataset_union.ipynb <==
# Unindo os datasets finais, com as labels

Após rodar o modelo para as ratings de 3 e 4, e setar as labels dos datasets com ratings 1, 2 e 5, ficamos com 2 datasets, que precisam ser unidos.

import pandas as pd

# Carregando os datasets
df_reviews = pd.read_csv("drive/My Drive/Colab Notebooks/TCC/dataset_with_labels_and_tokens.csv", index_col=0)
df_reviews_rest = pd.read_csv("drive/My Drive/Colab Notebooks/TCC/without_label_labeled_bert.csv", index_col=0)

# Verificando se o shape está correto
print(df_reviews.shape, df_reviews_rest.shape)

# União dos datasets
df_final = pd.concat([df_reviews, df_reviews_rest])
df_final = df_final.reset_index()

# Conferindo os indices
print(len(df_reviews.index))
print(len(df_reviews_rest.index))
print(len(df_final.index))

# Gerando dataset final
df_final.to_csv("drive/My Drive/Colab Notebooks/TCC/dataset_completed.csv",index=False)
